[POC] Implement toolkit for project name prediction ---------------------------------------------------
The POC implementation of a trainable classifier makes use of Naive
Bayes classifier. An approach to predict project names using integrated
pipeline is suggested and in more depth will be described in a
self-containing repository
[nvd-toolkit](https://github.com/fabric8-analytics/fabric8-analytics-nvd-toolkit).

The implementation relys on NLTK text processing library and sklearn
pipelines.

As of evaluation and cross validation, it is yet to be estimated more
precisely in the future work.

Implement FeatureExtractor preprocessor

- the extractor implements `transform` method and can therefore be used
in scikit pipelines along with other preprocessors and classifiers
- FeatureExtractor takes as an argument output of NLTKPreprocessor
- the extractor is extensible with hooks defined by Hook class

Implement GitHub handler and NVDFeedPreprocessor

- the NVDFeedPreprocessor is reponsible for NVD data preprocessing, that
is getting relevant data for each cve, filtering the feed for desired
cves
- to handle the cves, it makes use of GitHubHandler which serves to
query the GitHub API and infereing information about the projects. This
is a basic handler which will be used for now.

- the preprocessor will provide data to label encoder, which is yet to
be implemented

Implement LabelPreprocessor

- LabelPreprocessor is responsible for extracting the data used in the
training pipeline and labeling them using a custom Hook

- The preprocessing pipeline should now be ready for use

Signed-off-by: Marek Cermak <macermak@redhat.com>

Initial implementation of NBClassifier

- the core part of classification and the final part of the pipeline is
the classifier
- the classifier can be trained and make predictions of the data
- it is possible (and is the common case) to export the trained
classifier and restore it to make predictions

- pipeline changes
- refactorings

Signed-off-by: Marek Cermak <macermak@redhat.com>